refactor: game module
